+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- MQTT status messages with statistics are now sent on every received file. The total
+  number of files received and the number of triggers actually fired are included in
+  the message payload. Resolves [issue 146](https://github.com/danecreekphotography/node-deepstackai-trigger/issues/146).
 - Fix typos in the source code. Resolves [issue 170](https://github.com/danecreekphotography/node-deepstackai-trigger/issues/170).
 - Add a clear message after initialization indicating whether startup was successful.
   If it wasn't there's now a link to a troubleshooting wiki page for assistance. Resolves